pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kg-cli"
version = "0.1.0"
description = "Knowledge Graph CLI - 知识图谱管理工具"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "黄智泓" },
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1",
    "anthropic>=0.39",
    "sentence-transformers>=3.0",
    "qdrant-client>=1.12",
    "surrealdb>=1.0",
    "rich>=13.7",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.9",
    "pymupdf>=1.24",
    "feedparser>=6.0",
    "trafilatura>=1.12",
    "html2text>=2024.2",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.8",
    "ruff>=0.5",
]

[project.scripts]
kg = "agents.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["agents"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
